svc/cortex/pyproject.toml
[project]
name = "cortexsvc"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.2.1",
    "pydantic==2.12.3",
    "pydantic-settings>=2.10.1",
    "qbrixcore==0.1.0a0",
    "qbrixstore==0.1.0a0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pre-commit>=4.0.1",
    "black>=24.0.0",
    "pytest-mock>=3.14.1",
]

[tool.uv]
prerelease = "if-necessary-or-explicit"

[tool.uv.sources]
qbrixcore = { index = "envelope" }
qbrixstore = { index = "envelope" }

[[tool.uv.index]]
name = "envelope"
url = "https://app.envelope.dev/simple"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=cortexsvc",
    "--cov-report=term-missing",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[project.scripts]
server = "cortexsvc.cli:run"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"